hw: skeleton for building on x86_32 and x86_64
This patch contains the initial code needed to build and bootstrap the
base-hw kernel on x86 platforms. On x86_32, the kernel gets as far as
printing messages via comport 1. On x86_64, it gets stuck earlier
because the binary contains 64-bit instructions, but it is started in
32-bit mode. The initial setup of page tables and switch to long mode is
still missing from the x86_64 crt0 code.

There are still some ARM-related bits spilled thoughout the x86-specific
code. @m-stein and @skalk: Do you see ways to improve it?