hw: Add syscall helper macros
The macros are used to assign syscall arguments to specific registers.
Using the AMD64 parameter passing convention avoids additional copying of
variables since the C++ function parameters are already in the right
registers.

Note that the rcx register cannot be used for argument 5 since the
syscall instruction saves the userspace rip in that register.