vbox: Force refresh of display in VGA device model
On Muen there is no possibility to map or unmap guest memory. Thus the
dirty page tracking algorithm of the VGA device model cannot be used.
Map the framebuffer memory unconditionally and always update the entire
display in the VGA device model.

This change substantially speeds up execution of the graphics operations
in Tinycore Linux with X server.