vbox: Enable configurable full refresh of VGA display
The patch allows to force a full display update in the VGA device model
via setting the following ExtraDataItem in the .vbox file:

  <ExtraDataItem name="VBoxInternal/Devices/vga/0/Config/FullDisplayUpdate" value="1"/>

This is required for Virtualbox on hw_x86_64_muen as there is no
possibility to map or unmap guest memory during runtime. Thus the dirty
page tracking algorithm of the VGA device model cannot be used.