vbox: Inject pending guest VM interrupts on Muen
Use mapped subject pending interrupts page of guest VM to perform
interrupt injection. IRQs are transferred from the Virtualbox trap
manager state to the pending interrupts region for injection. If an IRQ
remains pending upon returning to the recompiler, it is copied back to
the trap manager state and cleared in the subject interrupts region.